% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{Alcohol}
\alias{Alcohol}
\title{Alcohol Consumption per Capita}
\format{
A data frame with 411 observations on the following variables.
\describe{
\item{\code{country}} {country name}
\item{\code{year}} {year}
\item{\code{alcohol}} {estimated per capita alcohol consumption for adults (15+)
in litres pure alcohol}
}
}
\source{
Gapminder (\url{https://www.gapminder.org/data/})
}
\usage{
data(Alcohol)
}
\description{
These data provide per capita alcohol consumption values
for many countries in 2005 and 2008.  There are also a few
countries for which there are data in other years.
}
\examples{
data(Alcohol)
# There are only a few observations in years other than 2005 and 2008
subset(Alcohol, ! year \%in\% c(2005,2008))
}
