% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirthdaysDoc.R, R/datasetsDoc.R
\docType{data}
\name{Birthdays}
\alias{Birthdays}
\title{US Birthdays}
\format{A data.frame with 7305 observations on the following 8 variables.
\itemize{
   item{code{date }}{[Date]}
   item{code{births }}{number of births on \code{date} [integer]}
   item{code{wday }}{day of week [ordered factor]}
   item{code{year }}{year [integer]}
   item{code{month }}{month [integer]}
   item{code{day }}{day of month [integer]}
   item{code{day_of_year }}{day of year [integer]}
   item{code{day_of_week }}{day of week [integer]}
}}
\source{
Data source: National Vital Statistics System natality data, as provided by
Google BigQuery and exported to csv 
Robert Kern (\url{http://www.mechanicalkern.com/static/birthdates-1968-1988.csv)}
}
\usage{
data(Birthdays)

data(Birthdays)
}
\description{
Number of births each day from 1968 to 1988

A day by day record of the number of births in each US State.
}
\details{
The number of births in \code{\link{Births78}} is slightly lower than the number
of births in this data set for the year 1978.  See the examples.
}
\examples{
data(Birthdays)
if(require(ggplot2)) {
  ggplot(data = Birthdays, aes(x = date, y = births, colour = wday)) +
    stat_smooth(se = FALSE, alpha = 0.8, geom = "line")
  ggplot(data = Birthdays, aes(x = day_of_year, y = births, colour = wday)) +
    geom_point(size = 0.4, alpha = 0.5) +
    stat_smooth(se = FALSE, geom = "line", alpha = 0.6, size = 1.5)
  if (require(dplyr)) {
    ggplot(
     data =  bind_cols(Birthdays \%>\% filter(year == 1978), 
                       Births78 \%>\% rename(births78 = births)),
     aes(x = births - births78)
     ) +
     geom_histogram(binwidth = 1)
  }
}

data(Birthdays)
if (require(lattice)) {
  xyplot(births ~ date, Birthdays, subset=state=="CA")
  xyplot(births ~ date, Birthdays, subset=state=="CA", 
    groups=wday, type='l')
  if (require(mosaic)) {
    xyplot(births ~ date, type='l',
      data = Birthdays \%>\% group_by(date) \%>\% summarise(births=sum(births)))
    }
  }
}

