% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirthsDoc.R
\docType{data}
\name{Births78}
\alias{Births78}
\title{US Births in 1978}
\format{
A data.frame with 365 observations on the following 8 variables.
\itemize{
\item \code{date}   Date
\item \code{births} Number of births on \code{date} (integer)
\item \code{wday}   Day of week (ordered factor)
\item \code{year}   Year (integer)
\item \code{month}  Month (integer)
\item \code{day_of_year}  Day of year (integer)
\item \code{day_of_month} Day of month (integer)
\item \code{day_of_week}  Day of week (integer)
}
}
\usage{
data(Births78)
}
\description{
A day by day record of the number of births in the United States in 1978.
}
\examples{
data(Births78)
if (require(ggformula)) {
  gf_point(births ~ date, data = Births78)
  gf_point(births ~ date, data = Births78, color = ~ wday)
}
}
\seealso{
\code{\link{Births2015}}, \code{\link{Births}}
}
