% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Dimes.R
\docType{data}
\name{Dimes}
\alias{Dimes}
\title{Weight of dimes}
\format{A data frame with 30 observations on the following 2 variables.
\itemize{
\item{\code{mass}}{ mass of dime in grams}
\item{\code{year}}{ year the dime was minted}
}}
\source{
Data were collected by Michael Stob.
}
\usage{
data(Dimes)
}
\description{
Weights of a sample of dimes
}

