% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaratogaHouses.R
\docType{data}
\name{SaratogaHouses}
\alias{SaratogaHouses}
\title{Houses in Saratoga (2006)}
\source{
Data were originally collected by Candice Corvetti and Dick De Veaux 
(Williams College) from \url{http://www.saratogacountyny.gov/departments/real-property-tax-service-agency/}.
}
\usage{
data(SaratogaHouses)
}
\description{
Data on houses in Saratoga, Florida
}
\details{
House characteristics in Saratoga, NY collected in 2006
}

