% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaratogaHouses.R
\docType{data}
\name{SaratogaHouses}
\alias{SaratogaHouses}
\title{Houses in Saratoga (2006)}
\format{A data frame with 1728 observations on the following 16 variables.
\itemize{
\item{\code{price}}{ price (1000s of US dollars)}
\item{\code{lotSize}}{ size of lot (square feet)}
\item{\code{age}}{ age of house (years)}
\item{\code{landValue}}{ value of land (1000s of US dollars)}
\item{\code{livingArea}}{ living are (square feet)}
\item{\code{pctCollege}}{ percent of neighborhood that graduated college}
\item{\code{bedrooms}}{ number of bedrooms}
\item{\code{firplaces}}{ number of fireplaces}
\item{\code{bathrooms}}{ number of bathrooms (half bathrooms have no shower or tub)}
\item{\code{rooms}}{ number of rooms}
\item{\code{heating}}{ type of heating system}
\item{\code{fuel}}{ fuel used for heating}
\item{\code{sewer}}{ type of sewer system}
\item{\code{waterfront}}{ whether property includes waterfront}
\item{\code{newConstruction}}{ whether the property is a new construction}
\item{\code{centralAir}}{ whether the house has central air}
}}
\source{
Data were originally collected by Candice Corvetti and Dick De Veaux 
(Williams College) from \url{http://www.saratogacountyny.gov/departments/real-property-tax-service-agency/}.
}
\usage{
data(SaratogaHouses)
}
\description{
Data on houses in Saratoga, Florida
}
\details{
House characteristics in Saratoga, NY collected in 2006
}

